SI-4694 Updated the API to use uint64_t types instead of intptr_t and uint64_t

+++ include/vmem/vmem_fram.h
@@ -32,8 +32,8 @@ typedef struct {
 		.ack_with_pull = 1, \
 	};
 
-void vmem_fram_read(vmem_t * vmem, uint64_t addr, void * dataout, intptr_t len);
-void vmem_fram_write(vmem_t * vmem, uint64_t addr, const void * datain, intptr_t len);
+void vmem_fram_read(vmem_t * vmem, uint64_t addr, void * dataout, uint64_t len);
+void vmem_fram_write(vmem_t * vmem, uint64_t addr, const void * datain, uint64_t len);
 
 
 #endif /* SRC_PARAM_VMEM_FRAM_H_ */
