+++ src/vmem/vmem_fram.c
@@ -18,11 +18,11 @@ void fram_write_data(uint32_t addr, const void *data, uint32_t len);
 void fram_read_data(uint32_t addr, void *data, uint32_t len);
 
 
-void vmem_fram_read(vmem_t * vmem, uint64_t addr, void * dataout, intptr_t len) {
+void vmem_fram_read(vmem_t * vmem, uint64_t addr, void * dataout, uint64_t len) {
 	fram_read_data(((intptr_t) ((vmem_fram_driver_t*) vmem->driver)->fram_addr) + addr, dataout, len);
 }
 
-void vmem_fram_write(vmem_t * vmem, uint64_t addr, const void * datain, intptr_t len) {
+void vmem_fram_write(vmem_t * vmem, uint64_t addr, const void * datain, uint64_t len) {
 	fram_write_data(((intptr_t) ((vmem_fram_driver_t*) vmem->driver)->fram_addr) + addr, datain, len);
 }
